Restructuring frontend component organization and naming for better maintainability
