+++ theplotpot-frontend/src/App.js
@@ -5,38 +5,45 @@ import {
 } from 'react-router-dom'
 import { Spin } from 'antd'
 import Cookies from 'js-cookie'
-import Navbar from './components/Navbar'
-import Footer from './components/Footer'
-import Notification from './components/Notification'
-import UserPage from './pages/users/UserPage'
+import Navbar from './components/layout/Navbar'
+import Footer from './components/layout/Footer'
+import Notification from './components/utilities/Notification'
 import './App.css'
-import InfoModal from './components/InfoModal'
-import DarkModeToggle from './components/DarkModeToggle'
+import InfoModal from './components/modals/InfoModal'
+import DarkModeToggle from './components/utilities/DarkModeToggle'
 
-//Lazy loaded components
-const Home = lazy(() => import('./pages/home/Home'))
+// Lazy loaded components
 const Login = lazy(() => import('./pages/auth/Login'))
 const Signup = lazy(() => import('./pages/auth/Signup'))
+
 const StoryForm = lazy(() => import('./pages/newstory/StoryForm'))
-const StoryPage = lazy(() => import('./pages/storypage/StoryPage'))
 const AddChapter = lazy(() => import('./pages/newstory/AddChapter'))
-const StoriesPage = lazy(() => import('./pages/stories/StoriesPage'))
-const UserSettings = lazy(() => import('./pages/users/UserSettings'))
-const CookieConsentPopup = lazy(() => import('./components/CookieConsentPopup'))
-const ActivateAccount = lazy(() => import('./pages/auth/ActivateAccount'))
+
+const StoryPage = lazy(() => import('./pages/storypage/'))
+
 const ProtectedRoute = lazy(() => import('./pages/admin/ProtectedRoute'))
 const UserList = lazy(() => import('./pages/admin/UserList'))
-const NotFoundPage = lazy(() => import('./pages/notfound/NotFoundPage'))
 
+const UserPage = lazy(() => import('./pages/users/UserPage'))
+const UserSettings = lazy(() => import('./pages/users/UserSettings'))
+
+const Home = lazy(() => import('./pages/home'))
+const StoriesPage = lazy(() => import('./pages/stories'))
+
+const NotFoundPage = lazy(() => import('./pages/notfound'))
+
+const CookieConsentPopup = lazy(() => import('./components/modals/CookieConsentPopup'))
 
 const App = () => {
   const [showCookiePopup, setShowCookiePopup] = useState(false)
   const [showInfo, setShowInfo] = useState(false)
+
   useEffect(() => {
     if (!Cookies.get('cookieConsent')) {
       setShowCookiePopup(true)
     }
   }, [])
+
   const handleConsent = () => {
     Cookies.set('cookieConsent', 'true', { expires: 365 })
     setShowCookiePopup(false)
@@ -45,12 +52,10 @@ const App = () => {
     }, 5000)
   }
 
-
   return (
-
     <div id="app">
       <Router>
-        <Navbar/>
+        <Navbar />
         <div className="main-content">
           <Suspense fallback={<div style={{ display: 'flex', justifyContent: 'center', alignItems: 'center', height: '80vh' }}><Spin size="large" /></div>}>
             <Routes>
@@ -64,7 +69,6 @@ const App = () => {
               <Route path="/add-chapter" element={<AddChapter />} />
               <Route path="/usersettings" element={<UserSettings />} />
               <Route path="/stories" element={<StoriesPage />} />
-              <Route path="/activate" element={<ActivateAccount />} />
               <Route
                 path="/admin/users"
                 element={
@@ -85,12 +89,10 @@ const App = () => {
         </Suspense>
       )}
       {showInfo && <InfoModal />}
-      <Footer></Footer>
+      <Footer />
       <DarkModeToggle />
     </div>
-
   )
 }
 
 export default App
-
