+++ theplotpot-frontend/src/components/layout/Navbar/index.js
@@ -1,0 +1,1 @@
+export { default } from './Navbar'
