+++ theplotpot-frontend/src/components/modals/CookieConsentPopup/index.js
@@ -1,0 +1,1 @@
+export { default } from './CookieConsentPopup'
