+++ theplotpot-frontend/src/components/modals/InfoModal/InfoModal.jsx
