+++ theplotpot-frontend/src/components/modals/InfoModal/index.js
@@ -1,0 +1,1 @@
+export { default } from './InfoModal'
