+++ theplotpot-frontend/src/components/modals/PrivacyPolicy/index.js
@@ -1,0 +1,1 @@
+export { default } from './PrivacyPolicy'
