+++ theplotpot-frontend/src/components/utilities/DarkModeToggle/index.js
@@ -1,0 +1,1 @@
+export { default } from './DarkModeToggle'
