+++ theplotpot-frontend/src/components/utilities/Notification/index.js
@@ -1,0 +1,1 @@
+export { default } from './Notification'
