+++ theplotpot-frontend/src/index.js
@@ -2,9 +2,9 @@ import React from 'react'
 import ReactDOM from 'react-dom/client'
 import App from './App'
 import { ApolloProvider, ApolloClient, InMemoryCache, HttpLink } from '@apollo/client'
-import { AuthProvider } from './pages/auth/AuthContext'
-import { NotificationsProvider } from './components/NotificationsContext'
-import { DarkModeProvider } from './components/DarkModeContext'
+import { AuthProvider } from './context/AuthContext'
+import { NotificationsProvider } from './context/NotificationsContext'
+import { DarkModeProvider } from './context/DarkModeContext'
 import { setContext } from '@apollo/client/link/context'
 import 'bootstrap/dist/css/bootstrap.css'
