+++ theplotpot-frontend/src/pages/admin/UserList.jsx
@@ -4,7 +4,7 @@ import { Table, Button } from 'react-bootstrap'
 import { Popconfirm } from 'antd'
 import { GET_ALL_USERS, BAN_USER, UNBAN_USER, DELETE_USER } from '../../api/queries'
 import { Link } from 'react-router-dom'
-import { useDarkMode } from '../../components/DarkModeContext'
+import { useDarkMode } from '../../context/DarkModeContext'
 
 const UserList = () => {
   const { isDarkMode } = useDarkMode()
