+++ theplotpot-frontend/src/pages/admin/index.js
@@ -1,0 +1,2 @@
+export { default as ProtectedRoute } from './ProtectedRoute'
+export { default as UserList } from './UserList'
