+++ theplotpot-frontend/src/pages/notfound/index.js
@@ -1,0 +1,1 @@
+export { default } from './NotFoundPage'
