+++ theplotpot-frontend/src/pages/stories/index.js
@@ -1,0 +1,1 @@
+export { default } from './StoriesPage'
