+++ theplotpot-frontend/src/pages/storypage/Chapter.jsx
@@ -1,17 +1,17 @@
 import { useState, useEffect } from 'react'
 import DOMPurify from 'dompurify'
-import './chapter.css'
+import './Chapter.css'
 import { Button } from 'react-bootstrap'
-import { useAuth } from '../auth/AuthContext'
+import { useAuth } from '../../context/AuthContext'
 import { ArrowLeft, HandThumbsUpFill } from 'react-bootstrap-icons'
-import { useLikeUnlikeChapter, useIsChapterLiked } from '../../hooks/likes'
+import { useLikeUnlikeChapter, useIsChapterLiked } from '../../hooks/useLikes.js'
 import Comments from './Comments'
 import { Popconfirm } from 'antd'
 import { useMutation } from '@apollo/client'
 import { DELETE_CHAPTER, DELETE_STORY } from '../../api/queries'
 import { useNavigate } from 'react-router-dom'
-import { useNotifications } from '../../components/NotificationsContext'
-import { useDarkMode } from '../../components/DarkModeContext'
+import { useNotifications } from '../../context/NotificationsContext'
+import { useDarkMode } from '../../context/DarkModeContext'
 import '../../utils/theme.css'
 
 const Chapter = ({ chapter, childChapters, onNavigate, onAddChapter, onGoBack, isLoading }) => {
