+++ theplotpot-frontend/src/pages/storypage/index.js
@@ -1,0 +1,1 @@
+export { default } from './StoryPage'
